Support util class generation in plugin
If no util methods are required, then util class will not be generated. Otherwise, we will generate an util class with or without Mockito usage, depending on whether Mockito is used in the tests or not. Note that whenever an util class with Mockito is required, it will be generated and it will rewrite a previous utils file if it exists. That's because the previous file could have been without Mockito.