Update util methods to work with fields of exact classes
Instead of searching the class hierarchy bottom to top until a field with a specified name is found we now search the field in its actual declaring class directly. That means that now we will not be able to access the wrong field by stopping too soon in the bottom to top search as we could before.